se cambia la logica de servicio, se elimina modelo examen, se crea modelo respuesta

+++ back/src/main/java/pep3/savingsemester/back/Entities/PreguntaE.java
@@ -17,5 +17,5 @@ public class PreguntaE {
     public Integer numPregunta;
     public Integer dificultad;
     public String link;
-    public Integer respuesta;
+    public Double respuesta;
 }
